+++ app/src/main/java/com/concordium/wallet/data/walletconnect/SignMessageParams.kt
@@ -1,13 +1,36 @@
 package com.concordium.wallet.data.walletconnect
 
 import com.concordium.wallet.App
+import com.google.gson.JsonSyntaxException
 
-data class SignMessageParams(
-    val message: String,
-) {
+sealed interface SignMessageParams {
+    val data: String
+
+    class Text(
+        override val data: String,
+    ) : SignMessageParams
+
+    class Binary(
+        val schema: Schema,
+        /**
+         * Hex-encoded binary data.
+         */
+        override val data: String,
+    ): SignMessageParams
+
     companion object {
-        fun fromSessionRequestParams(signMessageParams: String): SignMessageParams {
-            return App.appCore.gson.fromJson(signMessageParams, SignMessageParams::class.java)
+        private val gson by lazy {
+            App.appCore.gson.newBuilder()
+                .registerTypeAdapter(SignMessageParams::class.java, SignMessageParamsDeserializer())
+                .create()
         }
+
+        /**
+         * @return parsed [SignMessageParams]
+         *
+         * @throws JsonSyntaxException if the request params can't be parsed
+         */
+        fun fromSessionRequestParams(sessionRequestParams: String): SignMessageParams =
+            gson.fromJson(sessionRequestParams, SignMessageParams::class.java)
     }
 }